requirements.txt
asgiref==3.5.2
Django==4.1.4
django-cors-headers==3.13.0
django-ipware==4.0.2
djangorestframework==3.14.0
drf-recaptcha==2.2.2
gunicorn==20.1.0
Pillow==9.3.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
python-magic==0.4.27
pytz==2022.6
sqlparse==0.4.3
